Quite bit of edits here.
To start I removed the function wrapping my answer button clicks and this solved my issue of the function running multiple times.

Next I let my setInterval = to a variable and this allowed me to stop it when I reached the end of the quiz.

Finally I made a retakeTest function that allows the user to retake the test if they decide they want to.